--- app/src/main/java/com/example/app_wordman/MainActivity.java
@@ -1,107 +1,0 @@
-package com.example.app_wordman;
-
-import androidx.appcompat.app.AlertDialog;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.viewpager.widget.ViewPager;
-
-import android.content.Context;
-import android.content.DialogInterface;
-import android.net.ConnectivityManager;
-import android.net.Network;
-import android.net.NetworkInfo;
-import android.net.wifi.p2p.WifiP2pManager;
-import android.os.Bundle;
-import android.util.Log;
-import android.view.View;
-import android.widget.Button;
-import android.widget.Toast;
-
-import com.example.app_wordman.adapter.PageAdapter;
-import com.google.android.material.tabs.TabItem;
-import com.google.android.material.tabs.TabLayout;
-
-public class MainActivity extends AppCompatActivity {
-
-    private TabLayout tabLayout;
-    private ViewPager viewPager;
-    private TabItem wordSearch, favWord;
-
-    private PageAdapter pageAdapter;
-
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        if(!isNetworkAvailable()==true) {
-            new AlertDialog.Builder(this)
-                    .setIcon(android.R.drawable.ic_dialog_alert)
-                    .setTitle("Internet Connection Alert")
-                    .setMessage("Please Check Your Internet Connection")
-                    .setPositiveButton("Close", new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialogInterface, int i) {
-                            finish();
-                        }
-                    }).show();
-        }
-
-
-        tabLayout = findViewById(R.id.tabLayout);
-        wordSearch = findViewById(R.id.tab1);
-        favWord= findViewById(R.id.tab2);
-
-        viewPager = findViewById(R.id.viewPager);
-
-        pageAdapter = new PageAdapter(getSupportFragmentManager(),tabLayout.getTabCount());
-        viewPager.setAdapter(pageAdapter);
-
-        tabLayout.getTabAt(1).setIcon(R.drawable.ic_baseline_star_24);
-        tabLayout.getTabAt(0).setIcon(R.drawable.ic_baseline_search_24);
-
-        tabLayout.setOnTabSelectedListener(new TabLayout.OnTabSelectedListener() {
-            @Override
-            public void onTabSelected(TabLayout.Tab tab) {
-                viewPager.setCurrentItem(tab.getPosition());
-                if(tab.getPosition() == 0){
-                    pageAdapter.notifyDataSetChanged();
-                }else if(tab.getPosition() == 1){
-                    pageAdapter.notifyDataSetChanged();
-                }
-            }
-
-            @Override
-            public void onTabUnselected(TabLayout.Tab tab) {
-
-            }
-
-            @Override
-            public void onTabReselected(TabLayout.Tab tab) {
-
-            }
-        });
-
-        viewPager.addOnPageChangeListener(new TabLayout.TabLayoutOnPageChangeListener(tabLayout));
-
-
-    }
-
-    private boolean isNetworkAvailable(){
-
-        ConnectivityManager connectivityManager = (ConnectivityManager) getSystemService(this.CONNECTIVITY_SERVICE);
-
-        NetworkInfo wificonn = connectivityManager.getNetworkInfo(connectivityManager.TYPE_WIFI);
-        NetworkInfo mobileconn = connectivityManager.getNetworkInfo(connectivityManager.TYPE_MOBILE);
-
-        if((wificonn != null && wificonn.isConnected()) || mobileconn != null && mobileconn.isConnected()){
-            return true;
-        }else{
-            return false;
-        }
-    }
-
-
-
-
-}
