--- app/src/main/java/com/example/app_wordman/adapter/PageAdapter.java
@@ -1,44 +1,0 @@
-package com.example.app_wordman.adapter;
-
-import androidx.annotation.NonNull;
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentPagerAdapter;
-
-import com.example.app_wordman.tabs.FavWord;
-import com.example.app_wordman.tabs.WordSearch;
-
-public class PageAdapter extends FragmentPagerAdapter {
-
-    private int numberOfTabs;
-
-    public PageAdapter(FragmentManager fm, int numberOfTabs) {
-        super(fm);
-        this.numberOfTabs = numberOfTabs;
-    }
-
-    @Override
-    public Fragment getItem(int position) {
-        switch(position) {
-            case 0:
-                return new WordSearch();
-            case 1:
-                return new FavWord();
-            default:
-                return null;
-        }
-    }
-
-    @Override
-    public int getItemPosition(@NonNull Object object) {
-        return POSITION_NONE;
-    }
-
-    @Override
-    public int getCount() {
-        return numberOfTabs;
-    }
-
-
-
-}
