+++ app/src/main/java/com/rezwanul502/app_wordman/api/DefinitionInterface.java
@@ -1,0 +1,13 @@
+package com.rezwanul502.app_wordman.api;
+
+import java.util.List;
+
+import retrofit2.Call;
+import retrofit2.http.GET;
+import retrofit2.http.Path;
+
+public interface DefinitionInterface {
+
+    @GET("entries/en/{word}")
+    Call<List<Example>> getDefinitions(@Path("word") String word);
+}
