+++ app/src/main/java/com/rezwanul502/app_wordman/splashScreen/SplashScreenActivity.java
@@ -1,0 +1,18 @@
+package com.rezwanul502.app_wordman.splashScreen;
+
+import android.content.Intent;
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.rezwanul502.app_wordman.MainActivity;
+
+public class SplashScreenActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        startActivity(new Intent(SplashScreenActivity.this, MainActivity.class));
+        finish();
+    }
+}
